examples/email_matching/requirements.txt
python-dotenv
openai
monkey-patch.py